B2: split atoms-2d `shape` enum atom into 4 separate atoms (arrow/cube/diamond/gear)
Aligns 2D atoms 1:1 with 3D shapes/{arrow-3d,cube-3d,diamond-3d,gear-3d}. The
former `shape` atom was a single multi-kind enum (cube/sphere/diamond/arrow/
gear/cylinder) which prevented per-type direct lift mapping. Sphere kind and
cylinder kind are dropped — 3D side has no twin (sphere-fill is data-viz, not
decorative shape; cylinder has no 3D atom).

Composite atoms cube-grid / gear-cluster / puzzle-pieces RETAINED — these have
no direct 3D twin (3D puzzle-piece-3d is singular, cube-segmented-3d is a sliced
cube not an N³ grid). Per user decision: 2D library expresses LLM vocabulary;
3D lift LLM may decompose composites into low-level primitives. Backlog item
added for future 3D-side composite atoms.

Files:
  CREATE shapes/arrow.js shapes/cube.js shapes/diamond.js shapes/gear.js
  DELETE shapes/shape.js
  UPDATE registry.js     - 4 new entries, drop `shape`
  UPDATE compositor-api.js MODE_2D_ADDENDUM - split into "single primitives" +
                                              "composites" sections
  UPDATE test-atoms-2d-framework.mjs - 4 separate atom tests (was 1 enum test)
  UPDATE examples/atoms-2d-demo/index.html - 4 demo cards (was 6, sphere+cylinder
                                              demos dropped)

Tests: 82/82 pass
