+++ sdf-js/examples/atoms-2d-demo/index.html
@@ -204,25 +204,19 @@
       { title: 'Matrix — SWOT 2x2',
         size: { w: 500, h: 500 },
         atom: { type: 'matrix-grid', args: { rows: 2, cols: 2, title: 'SWOT Analysis', cells: [{ label: 'Strengths' }, { label: 'Weaknesses' }, { label: 'Opportunities' }, { label: 'Threats' }] } } },
-      // ---- Shape primitives (Phase 3) ----
-      { title: 'Shape — Cube (Product)',
+      // ---- Shape primitives (Phase 3 — split from `shape` enum to 4 atoms 1:1 with 3D) ----
+      { title: 'Cube — Product',
         size: { w: 280, h: 280 },
-        atom: { type: 'shape', args: { kind: 'cube', label: 'Product' } } },
-      { title: 'Shape — Sphere (Global)',
+        atom: { type: 'cube', args: { label: 'Product' } } },
+      { title: 'Diamond — Premium',
         size: { w: 280, h: 280 },
-        atom: { type: 'shape', args: { kind: 'sphere', label: 'Global', color: [60, 180, 100] } } },
-      { title: 'Shape — Diamond (Premium)',
+        atom: { type: 'diamond', args: { label: 'Premium', color: [180, 60, 180] } } },
+      { title: 'Arrow Up — Growth',
         size: { w: 280, h: 280 },
-        atom: { type: 'shape', args: { kind: 'diamond', label: 'Premium', color: [180, 60, 180] } } },
-      { title: 'Shape — Arrow Up Growth',
+        atom: { type: 'arrow', args: { direction: 'up', label: '+127% Growth', color: [40, 160, 100] } } },
+      { title: 'Gear — Process',
         size: { w: 280, h: 280 },
-        atom: { type: 'shape', args: { kind: 'arrow', direction: 'up', label: '+127% Growth', color: [40, 160, 100] } } },
-      { title: 'Shape — Gear (Process)',
-        size: { w: 280, h: 280 },
-        atom: { type: 'shape', args: { kind: 'gear', label: 'Process', color: [200, 130, 40] } } },
-      { title: 'Shape — Cylinder (Database)',
-        size: { w: 280, h: 280 },
-        atom: { type: 'shape', args: { kind: 'cylinder', label: 'Database', color: [60, 100, 200] } } },
+        atom: { type: 'gear', args: { label: 'Process', color: [200, 130, 40] } } },
       // ---- Icon badges (Phase 4) ----
       { title: 'Icon — Users (Engineering)',
         size: { w: 240, h: 240 },
