+++ sdf-js/scripts/test-atoms-2d-framework.mjs
@@ -822,22 +822,20 @@ console.log('\n--- pyramid atom ---');
   ok(!crashed, 'empty layers no crash');
 }
 
-// ----- shape atom (Phase 3) -----
-console.log('\n--- shape atom ---');
+// ----- shape atoms (Phase 3 — split from `shape` enum to 4 atoms aligned with 3D) -----
+console.log('\n--- shape atoms (arrow/cube/diamond/gear) ---');
 {
-  const spec = await getAtomSpec('shape');
-  ok(spec.type === 'shape', 'shape spec.type');
-  ok(spec.category === 'shapes', `shape category = ${spec.category}`);
-  ok(spec.args.kind.required, 'kind required');
+  const types = ['arrow', 'cube', 'diamond', 'gear'];
+  for (const type of types) {
+    const spec = await getAtomSpec(type);
+    ok(spec.type === type, `${type} spec.type`);
+    ok(spec.category === 'shapes', `${type} category = ${spec.category}`);
 
-  // Each kind renders without crash
-  const kinds = ['cube', 'sphere', 'diamond', 'arrow', 'gear', 'cylinder'];
-  for (const kind of kinds) {
     const recorded = [];
     const c = stubCtxWithEllipse(recorded);
     let crashed = false;
     try {
-      await renderAtom(c, 'shape', { kind, label: `Test ${kind}` }, 'pseudo3d', {
+      await renderAtom(c, type, { label: `Test ${type}` }, 'pseudo3d', {
         x: 0,
         y: 0,
         w: 240,
@@ -846,32 +844,18 @@ console.log('\n--- shape atom ---');
       });
     } catch (e) {
       crashed = true;
-      console.error(`  ${kind} error: ${e.message}`);
+      console.error(`  ${type} error: ${e.message}`);
     }
-    ok(!crashed, `kind="${kind}" renders without crash`);
-    ok(recorded.includes(`Test ${kind}`), `kind="${kind}" label rendered`);
+    ok(!crashed, `${type} renders without crash`);
+    ok(recorded.includes(`Test ${type}`), `${type} label rendered`);
   }
-
-  // Unknown kind falls back gracefully
-  const c2 = stubCtxWithEllipse([]);
-  let crashed = false;
-  try {
-    await renderAtom(c2, 'shape', { kind: 'nonexistent' }, 'pseudo3d', {
-      w: 240,
-      h: 240,
-      palette: { bg: [255, 255, 255], silhouetteColor: [0, 0, 0] },
-    });
-  } catch (e) {
-    crashed = true;
-  }
-  ok(!crashed, 'unknown kind falls back gracefully (defaults to cube)');
 
   // Arrow direction param
   const recA = [];
   await renderAtom(
     stubCtxWithEllipse(recA),
-    'shape',
-    { kind: 'arrow', direction: 'up', label: 'Up' },
+    'arrow',
+    { direction: 'up', label: 'Up' },
     'pseudo3d',
     { w: 240, h: 240, palette: { bg: [255, 255, 255], silhouetteColor: [0, 0, 0] } },
   );
